+++ fixtures/components.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Components
+	extend Phlex::Bucket
+
+	autoload :SayHi, "components/say_hi"
+end
