+++ fixtures/components/say_hi.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+class Components::SayHi < Phlex::HTML
+	def initialize(name, times: 1)
+		@name = name
+		@times = times
+	end
+
+	def template
+		article {
+			@times.times { h1 { "Hi #{@name}" } }
+			yield
+		}
+	end
+end
